+++ Sample/VoxeetSDK Sample/ConferenceViewController.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 import VoxeetSDK
+import AVKit
 
 /*
  *  MARK: - Conference class
@@ -18,13 +19,18 @@ class ConferenceViewController: UIViewController {
     @IBOutlet weak var conferenceIDLabel: UILabel!
     @IBOutlet weak var tableView: UITableView!
     @IBOutlet weak var broadcastMessageTextView: UITextView!
-    @IBOutlet weak var screenShareView: VideoRenderer!
-    @IBOutlet weak var ownCameraView: VideoRenderer!
+    @IBOutlet weak var screenShareView: VTVideoView!
+    @IBOutlet weak var ownCameraView: VTVideoView!
     @IBOutlet weak var ownCameraHandlerButton: UIButton!
     @IBOutlet weak var switchDeviceSpeakerButton: UIButton!
+    @IBOutlet weak var startScreenShareButton: UIButton!
+    @IBOutlet weak var videoPresentationView: UIView!
+    @IBOutlet weak var filePresentationImageView: UIImageView!
     
     // Current conference ID.
     var conferenceID: String?
+    
+    private var player: AVPlayer?
     
     /*
      *  MARK: Load / Unload
@@ -64,15 +70,41 @@ class ConferenceViewController: UIViewController {
             ownCameraHandlerButton.isHidden = true
         }
         
+        // Screen share feature is only available above (or equal to) iOS 11.
+        if #available(iOS 11.0, *) {} else {
+            startScreenShareButton.isHidden = true
+        }
+        
         // Select / deselect the switchDeviceSpeakerButton when an headset is plugged.
         NotificationCenter.default.addObserver(self, selector: #selector(audioSessionRouteChange), name: NSNotification.Name.AVAudioSessionRouteChange, object: nil)
+        
+        // Force the device screen to never going to sleep mode.
+        UIApplication.shared.isIdleTimerDisabled = true
+        
+        // File presentation observers.
+        NotificationCenter.default.addObserver(self, selector: #selector(filePresentationStartedUpdated), name: .VTFilePresentationStarted, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(filePresentationStartedUpdated), name: .VTFilePresentationUpdated, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(filePresentationStopped), name: .VTFilePresentationStopped, object: nil)
+        
+        // Video presentation observers.
+        NotificationCenter.default.addObserver(self, selector: #selector(videoPresentationStarted), name: .VTVideoPresentationStarted, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(videoPresentationStopped), name: .VTVideoPresentationStopped, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(videoPresentationPlay), name: .VTVideoPresentationPlay, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(videoPresentationPause), name: .VTVideoPresentationPause, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(videoPresentationSeek), name: .VTVideoPresentationSeek, object: nil)
     }
     
     deinit {
         // Debug.
         print("[DEBUG] \(#function)")
         
+        // Remove observers.
         NotificationCenter.default.removeObserver(self)
+        // Reset: Force the device screen to never going to sleep mode.
+        UIApplication.shared.isIdleTimerDisabled = false
+        
+        player?.pause()
+        player = nil
     }
     
     /*
@@ -107,6 +139,35 @@ class ConferenceViewController: UIViewController {
         alertController.addAction(confirmAction)
         alertController.addAction(cancelAction)
         self.present(alertController, animated: true, completion: nil)
+    }
+    
+    @IBAction func startScreenShare(_ sender: UIButton) {
+        guard VoxeetSDK.shared.conference.getScreenShareMediaStream() == nil else {
+            // Debug.
+            print("[ERROR] \(#function) - Error: Only one screen share allowed.")
+            return
+        }
+        
+        if #available(iOS 11.0, *) {
+            if sender.isSelected {
+                VoxeetSDK.shared.conference.startScreenShare(completion: { (error) in
+                    if let error = error {
+                        // Debug.
+                        print("[ERROR] \(#function) - Error: \(error)")
+                    }
+                })
+            } else {
+                VoxeetSDK.shared.conference.stopScreenShare(completion: { (error) in
+                    if let error = error {
+                        // Debug.
+                        print("[ERROR] \(#function) - Error: \(error)")
+                    }
+                })
+            }
+            
+            sender.isSelected = !sender.isSelected
+            sender.setTitle(sender.isSelected ? "Start screen share" : "Stop screen share", for: .normal)
+        }
     }
     
     @IBAction func switchDeviceSpeaker(_ button: UIButton) {
@@ -162,65 +223,56 @@ class ConferenceViewController: UIViewController {
  */
 
 extension ConferenceViewController: VTConferenceDelegate {
-    func participantAdded(userID: String, userInfo: [String: Any], stream: MediaStream) {
+    func participantJoined(userID: String, stream: RTCMediaStream) {
         if VoxeetSDK.shared.session.user?.id == userID {
             // Attaching own user's video stream.
-            ownCameraView.isHidden = false
-            VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: ownCameraView)
-        } else {
-            // Save the user and refresh the tableView.
-            tableView.reloadData()
-            
-            // Getting user's cell.
-            let users = VoxeetSDK.shared.conference.users
-            if let index = users.index(where: { $0.id == userID }), let cell = self.tableView.cellForRow(at: IndexPath(row: index, section: 0)) as? ConferenceTableViewCell {
-                // Attaching user's video stream.
-                cell.userVideoView.isHidden = false
-                VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: cell.userVideoView)
+            if !stream.videoTracks.isEmpty {
+                ownCameraView.isHidden = false
+                VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: ownCameraView)
             }
+        } else {
+            tableView.reloadData()
         }
     }
     
-    func participantUpdated(userID: String, userInfo: [String: Any], stream: MediaStream) {
-        let users = VoxeetSDK.shared.conference.users
-        var renderer: VideoRenderer?
+    func participantUpdated(userID: String, stream: RTCMediaStream) {
+        let users = VoxeetSDK.shared.conference.users.filter({ $0.asStream })
         
         // Get the video renderer.
         if VoxeetSDK.shared.session.user?.id == userID {
-            renderer = ownCameraView
-        } else if let index = users.index(where: { $0.id == userID }), let cell = self.tableView.cellForRow(at: IndexPath(row: index, section: 0)) as? ConferenceTableViewCell {
-            renderer = cell.userVideoView
-            renderer?.isHidden = !stream.hasVideo
-        }
-        
-        // Attach / unattach the stream.
-        if let renderer = renderer {
-            if stream.hasVideo {
-                VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: renderer)
-            } else {
-                VoxeetSDK.shared.conference.unattachMediaStream(stream, renderer: renderer)
+            ownCameraView.isHidden = stream.videoTracks.isEmpty
+            
+            // Attaching own user's video stream.
+            if !stream.videoTracks.isEmpty {
+                VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: ownCameraView)
             }
+        } else if let index = users.index(where: { $0.id == userID }) {
+            tableView.reloadRows(at: [IndexPath(row: index, section: 0)], with: .none)
         }
     }
     
-    func participantRemoved(userID: String, userInfo: [String: Any]) {
+    func participantLeft(userID: String) {
         tableView.reloadData()
     }
     
-    func messageReceived(userID: String, userInfo: [String: Any], message: String) {
-        if let name = VoxeetSDK.shared.conference.user(userID: userID)?.externalName() {
+    func messageReceived(userID: String, message: String) {
+        if let name = VoxeetSDK.shared.conference.user(userID: userID)?.name {
             broadcastMessageTextView.text = "\(name): \(message)"
         } else {
             broadcastMessageTextView.text = "\(userID): \(message)"
         }
     }
     
-    func screenShareStarted(userID: String, stream: MediaStream) {
+    func screenShareStarted(userID: String, stream: RTCMediaStream) {
         // Attaching a video stream to a renderer.
         VoxeetSDK.shared.conference.attachMediaStream(stream, renderer: screenShareView)
+        
+        screenShareView.alpha = 1
     }
     
-    func screenShareStopped(userID: String) {}
+    func screenShareStopped(userID: String) {
+        screenShareView.alpha = 0
+    }
 }
 
 /*
@@ -229,14 +281,14 @@ extension ConferenceViewController: VTConferenceDelegate {
 
 extension ConferenceViewController: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return VoxeetSDK.shared.conference.users.count
+        return VoxeetSDK.shared.conference.users.filter({ $0.asStream }).count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "tableViewCell", for: indexPath) as! ConferenceTableViewCell
         
         // Getting the current user.
-        let users = VoxeetSDK.shared.conference.users
+        let users = VoxeetSDK.shared.conference.users.filter({ $0.asStream })
         let user = users[(indexPath as NSIndexPath).row]
         
         // Setting up the cell.
@@ -248,9 +300,11 @@ extension ConferenceViewController: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
         
-        // Mute a user.
-        let users = VoxeetSDK.shared.conference.users
+        // Getting the current user.
+        let users = VoxeetSDK.shared.conference.users.filter({ $0.asStream })
         let user = users[(indexPath as NSIndexPath).row]
+        
+        // Mute a user.
         if let userID = user.id {
             let isMuted = VoxeetSDK.shared.conference.toggleMute(userID: userID)
             
@@ -259,5 +313,104 @@ extension ConferenceViewController: UITableViewDataSource, UITableViewDelegate {
                 cell.backgroundColor = isMuted ? UIColor.red : UIColor.white
             }
         }
+    }
+}
+
+/*
+ *  MARK: - Conference file presentation
+ */
+
+extension ConferenceViewController {
+    @objc func filePresentationStartedUpdated(notification: Notification) {
+        guard let userInfo = notification.userInfo?.values.first as? Data else {
+            return
+        }
+        
+        do {
+            if let json = try JSONSerialization.jsonObject(with: userInfo) as? [String: Any] {
+                if let fileID = json["fileId"] as? String, let page = json["position"] as? Int {
+                    if let url = VoxeetSDK.shared.filePresentation.getImage(fileID: fileID, page: page) {
+                        let data = try Data(contentsOf: url)
+                        let image = UIImage(data: data)
+                        filePresentationImageView.image = image
+                        filePresentationImageView.isHidden = false
+                    }
+                }
+            }
+        } catch {}
+    }
+    
+    @objc func filePresentationStopped(notification: Notification) {
+        filePresentationImageView.image = nil
+        filePresentationImageView.isHidden = true
+    }
+}
+
+/*
+ *  MARK: - Conference video presentation
+ */
+
+extension ConferenceViewController {
+    @objc func videoPresentationStarted(notification: Notification) {
+        guard let userInfo = notification.userInfo?.values.first as? Data else {
+            return
+        }
+        
+        do {
+            if let json = try JSONSerialization.jsonObject(with: userInfo) as? [String: Any] {
+                if let url = URL(string: json["url"] as? String ?? ""), let timestamp = json["timestamp"] as? Int {
+                    player = AVPlayer(url: url)
+                    let playerLayer = AVPlayerLayer(player: player)
+                    playerLayer.frame = videoPresentationView.bounds
+                    playerLayer.backgroundColor = UIColor.black.cgColor
+                    videoPresentationView.layer.addSublayer(playerLayer)
+                    player?.seek(to: CMTimeMakeWithSeconds(Double(timestamp) / 1000, 1000)) { _ in
+                        self.player?.play()
+                    }
+                    
+                    videoPresentationView.isHidden = false
+                }
+            }
+        } catch {}
+    }
+    
+    @objc func videoPresentationStopped(notification: Notification) {
+        player?.pause()
+        player = nil
+        videoPresentationView.isHidden = true
+    }
+    
+    @objc func videoPresentationPlay(notification: Notification) {
+        guard let userInfo = notification.userInfo?.values.first as? Data else {
+            return
+        }
+        
+        do {
+            if let json = try JSONSerialization.jsonObject(with: userInfo) as? [String: Any] {
+                if let timestamp = json["timestamp"] as? Int {
+                    player?.seek(to: CMTimeMakeWithSeconds(Double(timestamp) / 1000, 1000)) { _ in
+                        self.player?.play()
+                    }
+                }
+            }
+        } catch {}
+    }
+    
+    @objc func videoPresentationPause(notification: Notification) {
+        player?.pause()
+    }
+    
+    @objc func videoPresentationSeek(notification: Notification) {
+        guard let userInfo = notification.userInfo?.values.first as? Data else {
+            return
+        }
+        
+        do {
+            if let json = try JSONSerialization.jsonObject(with: userInfo) as? [String: Any] {
+                if let timestamp = json["timestamp"] as? Int {
+                    player?.seek(to: CMTimeMakeWithSeconds(Double(timestamp) / 1000, 1000))
+                }
+            }
+        } catch {}
     }
 }